pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

# ---------------------------------------------------------------------------
# Project metadata
# ---------------------------------------------------------------------------
[project]
name = "pytmb"
version = "1.6.0"
description = "Calculate a Tumour Mutational Burden (TMB) score from a VCF file."
readme = "README.md"
license = { file = "LICENSE" }
authors = [
    { name = "Tom Gutman",      email = "tom.gutman@curie.fr" },
    { name = "Nicolas Servant", email = "nicolas.servant@curie.fr" },
]
keywords = ["bioinformatics", "TMB", "tumour mutational burden", "VCF", "cancer genomics"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved",
    "Topic :: Scientific/Engineering :: Bio-Informatics",
    "Intended Audience :: Science/Research",
    "Operating System :: POSIX :: Linux",
]
requires-python = ">=3.9"

# Core runtime dependencies
dependencies = [
    "cyvcf2>=0.30",
    "pyyaml>=6.0",
    "numpy>=1.21",
    "scipy>=1.9",
]

[project.optional-dependencies]
# Extra dependencies needed only for pyEffGenomeSize
effgenomesize = [
    "pybedtools>=0.9",
    "pandas>=1.3",
]
# Development / testing extras
dev = [
    "pytest>=7",
    "pytest-cov",
]

# ---------------------------------------------------------------------------
# CLI entry points
# ---------------------------------------------------------------------------
[project.scripts]
pyTMB            = "pytmb.cli.run_tmb:main"
pyEffGenomeSize  = "pytmb.cli.run_effgenomesize:main"

# ---------------------------------------------------------------------------
# Package discovery and data files
# ---------------------------------------------------------------------------
[tool.setuptools.packages.find]
where = ["."]
include = ["pytmb*"]

[tool.setuptools.package-data]
pytmb = ["../config/*.yml"]

# ---------------------------------------------------------------------------
# pytest configuration
# ---------------------------------------------------------------------------
[tool.pytest.ini_options]
testpaths = ["test"]
